images/whisper-app/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
fastapi==0.115.5
filelock==3.16.1
fonttools==4.55.0
fsspec==2024.10.0
h11==0.14.0
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib==3.9.2
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.4.2
noisereduce==3.0.3
numba==0.60.0
numpy==2.0.2
openai-whisper==20240930
packaging==24.2
pillow==11.0.0
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
regex==2024.11.6
requests==2.32.3
scipy==1.14.1
semantic-version==2.10.0
setuptools==75.6.0
setuptools-rust==1.10.2
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.1
starlette==0.41.3
sympy==1.13.1
tiktoken==0.8.0
torch==2.5.1
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
fastapi==0.115.5
filelock==3.16.1
Flask==3.1.0
fonttools==4.55.0
fsspec==2024.10.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib==3.9.2
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.4.2
noisereduce==3.0.3
numba==0.60.0
numpy==2.0.2
openai-whisper==20240930
packaging==24.2
pillow==11.0.0
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
regex==2024.11.6
requests==2.32.3
scipy==1.14.1
semantic-version==2.10.0
setuptools==75.6.0
setuptools-rust==1.10.2
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.1
starlette==0.41.3
sympy==1.13.1
tiktoken==0.8.0
torch==2.5.1
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
Werkzeug==3.1.3
